+++ dnd-supplier-ms/src/main/java/com/capg/dnd/supplier/ms/service/SupplierService.java
@@ -2,15 +2,17 @@ package com.capg.dnd.supplier.ms.service;
 
 import java.util.List;
 
+
+
 import com.capg.dnd.supplier.ms.model.Supplier;
 
 public interface SupplierService{
 
     public Supplier addSupplier(Supplier suplier);
-	public boolean deleteSupplier(Long supplierId);
-	public Supplier getSupplierById(Long supplierId);
+	public boolean deleteSupplier(int supplierId);
+	public Supplier getSupplierById(int supplierId);
 			
 	public Supplier updateSupplier(Supplier supplier);
 	public List<Supplier> getAllSupplier();
-	
+
 }
